+++ versionsupport_1_11_R1/src/main/java/com/andrei1058/bedwars/support/version/v1_11_R1/Silverfish.java
@@ -41,7 +41,7 @@ public class Silverfish extends EntitySilverfish {
         this.targetSelector.a(2, new AttackEnemies<>(this, EntityHuman.class, true, bedWarsTeam, vs));
     }
 
-    public static LivingEntity spawn(VersionSupport vs, Location loc, ITeam team, int speed, int health, int despawn, int damage) {
+    public static LivingEntity spawn(VersionSupport vs, Location loc, ITeam team, double speed, double health, int despawn, double damage) {
         WorldServer mcWorld = ((CraftWorld)loc.getWorld()).getHandle();
         Silverfish customEnt = new Silverfish(mcWorld, team, vs);
         customEnt.setLocation(loc.getX(), loc.getY(), loc.getZ(), loc.getYaw(), loc.getPitch());
